@@ -5,11 +5,13 @@ import logging
 import time
 
 import matplotlib.pyplot as plt
-from IPython.display import display
-from matplotlib.ticker import MaxNLocator
 
-from cellpack_analysis.analysis.punctate_analysis import distance
-from cellpack_analysis.analysis.punctate_analysis.stats_functions import (
+from cellpack_analysis.analysis.punctate_analysis.lib import (
+    distance,
+    occupancy,
+    visualization,
+)
+from cellpack_analysis.analysis.punctate_analysis.lib.stats_functions import (
     normalize_distances,
 )
 from cellpack_analysis.lib.file_io import get_project_root
@@ -29,9 +31,8 @@ start_time = time.time()
 # SEC61B: ER
 # ST6GAL1: Golgi
 STRUCTURE_ID = "SLC25A17"
-PACKING_ID = "ER_peroxisome"
+PACKING_ID = "peroxisome"
 STRUCTURE_NAME = "peroxisome"
-STRUCT_RADIUS = 2.37  # 2.37 um for peroxisomes, 2.6 um for early endosomes
 # %% [markdown]
 # ### Set packing modes to analyze
 save_format = "svg"
@@ -41,15 +42,15 @@ packing_modes = [
     "nucleus_gradient_strong",
     "membrane_gradient_strong",
     # "apical_gradient",
-    "struct_gradient",
+    # "struct_gradient",
 ]
 
 channel_map = {
     "SLC25A17": "SLC25A17",
-    "random": "SEC61B",
-    "nucleus_gradient_strong": "SEC61B",
-    "membrane_gradient_strong": "SEC61B",
-    "struct_gradient": "SEC61B",
+    "random": "SLC25A17",
+    "nucleus_gradient_strong": "SLC25A17",
+    "membrane_gradient_strong": "SLC25A17",
+    # "struct_gradient": "SEC61B",
 }
 
 # relative path to packing outputs
@@ -100,7 +101,7 @@ all_positions = get_position_data_from_outputs(
     results_dir=results_dir,
     packing_output_folder=packing_output_folder,
     ingredient_key=f"membrane_interior_{STRUCTURE_NAME}",
-    recalculate=True,
+    recalculate=False,
 )
 # %% [markdown]
 # ### Get mesh information
@@ -109,7 +110,7 @@ for structure_id in all_structures:
     mesh_information_dict = get_mesh_information_dict_for_structure(
         structure_id=structure_id,
         base_datadir=base_datadir,
-        recalculate=True,
+        recalculate=False,
     )
     combined_mesh_information_dict[structure_id] = mesh_information_dict
 # %% [markdown]
@@ -120,10 +121,10 @@ all_distance_dict = distance.get_distance_dictionary(
     mesh_information_dict=combined_mesh_information_dict,
     channel_map=channel_map,
     results_dir=results_dir,
-    recalculate=True,
+    recalculate=False,
 )
 
-all_distance_dict = distance.filter_nans_from_distance_distribution_dict(
+all_distance_dict = distance.filter_invalids_from_distance_distribution_dict(
     distance_distribution_dict=all_distance_dict,
 )
 all_distance_dict = normalize_distances(
@@ -146,7 +147,7 @@ distance_kde_dict = distance.get_distance_distribution_kde(
     channel_map=channel_map,
     packing_modes=packing_modes,
     results_dir=results_dir,
-    recalculate=True,
+    recalculate=False,
     suffix=suffix,
     normalization=normalization,
     distance_measure=occupancy_distance_measure,
@@ -155,7 +156,7 @@ distance_kde_dict = distance.get_distance_distribution_kde(
 # %% [markdown]
 # ### Plot illustration for occupancy distribution
 kde_distance, kde_available_space, xvals, yvals, fig_ill, axs_ill = (
-    distance.plot_occupancy_illustration(
+    visualization.plot_occupancy_illustration(
         distance_dict=all_distance_dict[occupancy_distance_measure],
         kde_dict=distance_kde_dict,
         baseline_mode="random",
@@ -172,7 +173,7 @@ kde_distance, kde_available_space, xvals, yvals, fig_ill, axs_ill = (
 
 # %% [markdown]
 # ### Plot individual occupancy ratio
-figs_ind, axs_ind = distance.plot_individual_occupancy_ratio(
+figs_ind, axs_ind = visualization.plot_individual_occupancy_ratio(
     distance_dict=all_distance_dict[occupancy_distance_measure],
     kde_dict=distance_kde_dict,
     packing_modes=packing_modes,
@@ -188,7 +189,7 @@ figs_ind, axs_ind = distance.plot_individual_occupancy_ratio(
 )
 # %% [markdown]
 # ### plot mean and std of occupancy ratio
-figs_ci, axs_ci = distance.plot_mean_and_std_occupancy_ratio_kde(
+figs_ci, axs_ci = visualization.plot_mean_and_std_occupancy_ratio_kde(
     distance_dict=all_distance_dict[occupancy_distance_measure],
     kde_dict=distance_kde_dict,
     packing_modes=packing_modes,
@@ -209,13 +210,13 @@ figs_ci, axs_ci = distance.plot_mean_and_std_occupancy_ratio_kde(
 # fig
 # %% [markdown]
 # ### get combined space corrected kde
-combined_kde_dict = distance.get_combined_occupancy_kde(
+combined_kde_dict = occupancy.get_combined_occupancy_kde(
     all_distance_dict=all_distance_dict,
     mesh_information_dict=combined_mesh_information_dict,
     channel_map=channel_map,
     packing_modes=packing_modes,
     results_dir=results_dir,
-    recalculate=True,
+    recalculate=False,
     suffix=suffix,
     distance_measure=occupancy_distance_measure,
     # sample_size=10,
@@ -226,7 +227,7 @@ combined_kde_dict = distance.get_combined_occupancy_kde(
 occupancy_start_time = time.time()
 aspect = None
 method = "pdf"
-fig_combined, ax_combined = distance.plot_combined_occupancy_ratio(
+fig_combined, ax_combined = visualization.plot_combined_occupancy_ratio(
     combined_kde_dict=combined_kde_dict,
     packing_modes=packing_modes,
     suffix=suffix,
@@ -245,38 +246,8 @@ log.info(
     f"Time taken to plot occupancy ratio: {time.time() - occupancy_start_time:.2f} s"
 )
 # %% [markdown]
-# #### edit figure as needed
-ax = ax_combined
-fig = fig_combined
-
-aspect = 0.1
-ax.set_aspect(aspect)
-
-ax.xaxis.set_major_locator(MaxNLocator(5))
-ax.yaxis.set_major_locator(MaxNLocator(4))
-# ax.legend(loc="upper right", prop={"size": 14})
-# ax.legend().set_visible(False)
-ax.set_xlim((0, 6))
-ax.set_ylim((0, 3))
-ax.set_ylabel("Occupancy ratio")
-ax.set_xlabel("Distance from Nucleus \u03bcm")
-plt.tight_layout()
-display(fig)
-# %% [markdown]
-# #### save edited figure
-filename = (
-    figures_dir
-    / f"{occupancy_distance_measure}_combined_{method}_occupancy_ratio_edited{suffix}"
-)
-for format in ["png", "svg"]:
-    fig.savefig(
-        f"{filename}.{format}",
-        dpi=300,
-    )
-
-# %% [markdown]
 # ### plot binned occupancy ratio
-fig_binned, ax_binned = distance.plot_binned_occupancy_ratio(
+fig_binned, ax_binned = visualization.plot_binned_occupancy_ratio(
     distance_dict=all_distance_dict[occupancy_distance_measure],
     packing_modes=packing_modes,
     mesh_information_dict=combined_mesh_information_dict,
@@ -294,34 +265,34 @@ fig_binned, ax_binned = distance.plot_binned_occupancy_ratio(
 )
 # %% [markdown]
 # ### get EMD between occupied and available distances
-emd_occupancy_dict = distance.get_occupancy_emd(
+emd_occupancy_dict = occupancy.get_occupancy_emd(
     distance_dict=all_distance_dict[occupancy_distance_measure],
     kde_dict=distance_kde_dict,
     packing_modes=packing_modes,
     results_dir=results_dir,
-    recalculate=True,
+    recalculate=False,
     suffix=suffix,
 )
 # %% [markdown]
 # ### box and whisker plot for occupancy EMD values
-distance.plot_occupancy_emd_boxplot(
+visualization.plot_occupancy_emd_boxplot(
     emd_occupancy_dict=emd_occupancy_dict,
     suffix=suffix,
     figures_dir=figures_dir,
 )
 # %% [markdown]
 # ### run ks test for occupancy distributions
-ks_occupancy_dict = distance.get_occupancy_ks_test_dict(
+ks_occupancy_dict = occupancy.get_occupancy_ks_test_dict(
     distance_dict=all_distance_dict[occupancy_distance_measure],
     kde_dict=distance_kde_dict,
     packing_modes=packing_modes,
     results_dir=results_dir,
-    recalculate=True,
+    recalculate=False,
     suffix=suffix,
 )
 # %% [markdown]
 # ### plot ks test results
-distance.plot_occupancy_ks_test(
+visualization.plot_occupancy_ks_test(
     ks_occupancy_dict=ks_occupancy_dict,
     distance_measure=occupancy_distance_measure,
     figures_dir=figures_dir,
