@@ -255,7 +255,7 @@ def add_contour_to_axis(ax, projection, dim, domain_nuc, domain_mem, **kwargs):
     nuc_contour = skmeasure.find_contours(nuc_proj, 0.5)
     mem_contour = skmeasure.find_contours(mem_proj, 0.5)
 
-    for alias_cont, alias_color in zip([nuc_contour, mem_contour], ["cyan", "magenta"]):
+    for alias_cont, alias_color in zip([nuc_contour, mem_contour], ["cyan", "magenta"], strict=False):
         [
             ax.plot(c[:, 1], c[:, 0], lw=kwargs.get("lw", 1), color=alias_color)
             for c in alias_cont
@@ -687,7 +687,7 @@ class Projector:
         for alias, img in self.data.items():
             shape = img.shape
             pad = [int(0.5 * (self.box_size - s)) for s in shape]
-            pad = [(p, int(self.box_size - s - p)) for (s, p) in zip(shape, pad)]
+            pad = [(p, int(self.box_size - s - p)) for (s, p) in zip(shape, pad, strict=False)]
             if np.min([np.min([i, j]) for i, j in pad]) < 0:
                 raise ValueError(
                     f"Box of size {self.box_size} invalid for image of size: {shape}."
@@ -718,7 +718,7 @@ class Projector:
         ax.imshow(
             proj, cmap="inferno", origin="lower", vmin=self.gfp_vmin, vmax=self.gfp_vmax
         )
-        for alias_cont, alias_color in zip(["nuc", "mem"], ["cyan", "magenta"]):
+        for alias_cont, alias_color in zip(["nuc", "mem"], ["cyan", "magenta"], strict=False):
             [
                 ax.plot(c[:, 1], c[:, 0], lw=0.5, color=alias_color)
                 for c in contour[alias_cont]
