@@ -8,4 +8,4 @@ def cohens_d(x: np.ndarray, y: np.ndarray) -> float:
     pooled_std = np.sqrt(
         ((nx - 1) * np.var(x, ddof=1) + (ny - 1) * np.var(y, ddof=1)) / dof
     )
-    return np.abs((np.mean(x) - np.mean(y))) / pooled_std
+    return np.abs(np.mean(x) - np.mean(y)) / pooled_std
